+++ Changes.md
@@ -51,6 +51,7 @@ Breaking Changes
 - Box : Removed deprecated plug promotion methods. Use `PlugAlgo` instead.
 - CyclesLight : Removed `use_camera`, `use_diffuse`, `use_glossy`, `use_transmission`, `use_scatter`, and `lightgroup` parameter plugs as Cycles no longer considers these to be light parameters. Ray visibility and light group membership is now set via the `cycles:visibility:*` and `cycles:lightgroup` attributes on a CyclesAttributes node.
 - CyclesAttributes : Removed `cycles:shader:heterogeneous_volume` attribute as it is no longer used by Cycles.
+- CyclesOptions : Removed `cycles:background:volume_step_size` option as it is no longer used by Cycles.
 
 Build
 -----